requirements.txt
# CognEasy PoCs - FASE 2C
# Sistema integrado de análise médica com IA real
# Dependências principais e versões testadas

# Interface e Framework Web
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Inteligência Artificial
openai>=1.3.0
langchain>=0.0.350
langchain-community>=0.0.10

# Busca Semântica e Vetores
chromadb>=0.4.15
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Processamento de Dados
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Manipulação de Texto e NLP
nltk>=3.8.1
spacy>=3.7.0
python-dateutil>=2.8.2
regex>=2023.8.8

# Banco de Dados
sqlite3  # Built-in Python
sqlalchemy>=2.0.0

# Processamento de Arquivos
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pillow>=10.0.0

# Utilitários de Data e Tempo
pytz>=2023.3
arrow>=1.2.3

# HTTP e APIs
requests>=2.31.0
urllib3>=2.0.0

# Configuração e Ambiente
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Logging e Monitoramento
loguru>=0.7.0

# Visualização e Gráficos (Opcional)
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilitários Gerais
pathlib2>=2.3.7  # Para compatibilidade
tqdm>=4.65.0
rich>=13.5.0

# Validação e Testes
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Desenvolvimento (Opcional)
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Segurança
cryptography>=41.0.0
bcrypt>=4.0.0

# Performance
ujson>=5.8.0
orjson>=3.9.0

# Compatibilidade de Sistema
setuptools>=68.0.0
wheel>=0.41.0
pip>=23.2.0